change number type from Int to Double for matrix

+++ src/main/kotlin/com/hukuta94/pathnodebuilder/pipeline/filter/calculator/ComputeFullDistanceMatrixFilter.kt
@@ -10,7 +10,7 @@ class ComputeFullDistanceMatrixFilter : Function<DistanceMatrixData, DistanceMat
         val outputConnections = input.connections
 
         val distanceMatrix = Array(outputPositions.size) {
-            IntArray(
+            DoubleArray(
                 outputPositions.size
             )
         }
